feat(tests): reshaped test folder
