+++ tests/functionnal/environment_manager/mod.rs
@@ -1,0 +1,4 @@
+pub mod manager;
+pub mod types;
+mod utilities;
+pub use manager::EnvironmentManager;
