+++ tests/functionnal/environment_manager/utilities.rs
@@ -1,0 +1,96 @@
+use std::{path::Path, process::ExitStatus};
+
+use wormhole::network::ip::IpP;
+
+use crate::functionnal::environment_manager::types::{Service, CLI_BIN};
+
+/// Returns `true` if the given services runs a pod on the given network
+pub fn service_has_pod_on_network(service: &Service, network: &String) -> bool {
+    service
+        .pods
+        .iter()
+        .find(|(nw, _, _)| nw == network)
+        .is_some()
+}
+
+// Returns `true` if the service is matching the requirements
+pub fn service_filter(ip: &Option<IpP>, network: &Option<String>, service: &Service) -> bool {
+    network
+        .as_ref()
+        .map_or_else(|| true, |nw| service_has_pod_on_network(service, &nw))
+        && (ip.as_ref().map_or_else(|| true, |ip| service.ip == *ip))
+}
+
+/// Runs a command with the cli and returns it's stdout
+/// Returns (status, stdio, stderr)
+pub fn cli_command<I, S>(args: I) -> (ExitStatus, String, String)
+where
+    I: IntoIterator<Item = S> + std::fmt::Debug,
+    S: AsRef<std::ffi::OsStr>,
+{
+    log::trace!("Cli command with args {:?}", args);
+
+    let output = std::process::Command::new(CLI_BIN)
+        .args(args)
+        .output()
+        .expect("can't launch cli command");
+    (
+        output.status,
+        std::str::from_utf8(&output.stdout).unwrap().to_string(),
+        std::str::from_utf8(&output.stderr).unwrap().to_string(),
+    )
+}
+
+/// Cli commands to create a pod
+pub fn cli_pod_creation_command(
+    network_name: String,
+    service_ip: &IpP,
+    dir_path: &Path,
+    ip: &IpP,
+    connect_to: Option<&IpP>,
+) -> IpP {
+    let (status, _, _) = cli_command(&[
+        service_ip.to_string().as_ref(),
+        "template",
+        "-C",
+        dir_path.to_string_lossy().to_string().as_ref(),
+    ]);
+    assert!(status.success(), "template cli command failed");
+
+    let mut ip = ip.clone();
+
+    loop {
+        let (status, _, stderr) = cli_command({
+            let mut args = vec![
+                service_ip.to_string(),
+                "new".to_string(),
+                network_name.clone(),
+                "-C".to_string(),
+                dir_path.to_string_lossy().to_string(),
+                "-i".to_string(),
+                ip.to_string(),
+            ];
+
+            if let Some(peer) = connect_to {
+                args.push("-u".to_string());
+                args.push(peer.to_string());
+            }
+            args
+        });
+        if !status.success() && !stderr.contains("AddrInUse") {
+            log::error!("Cli stderr: {}", stderr);
+            assert!(status.success(), "'new' cli command failed");
+        } else if !status.success() {
+            assert!(
+                ip.get_ip_last() < 100,
+                "seems that 100+ pods are already running"
+            );
+            log::warn!("Pod ip {} already in use", ip);
+            ip.set_ip_last(ip.get_ip_last() + 1);
+        } else {
+            log::trace!("Created pod on {}", ip);
+            break;
+        }
+    }
+    ip
+}
