+++ tests/functionnal/test_sending_files_on_stop.rs
@@ -1,13 +1,18 @@
-use crate::functionnal::{append_to_path, environment_manager::types::SLEEP_TIME, start_log};
+use crate::functionnal::{append_to_path, environment_manager::SLEEP_TIME, start_log};
 
 use super::environment_manager;
 
 pub use environment_manager::EnvironmentManager;
 use serial_test::serial;
 
+/// Checks that when gracefully stopping a service, files exclusively on that service are sent to others
+/// - Creates two services with no redundancy
+/// - Creates a network
+/// - Stops the service that detains files
+/// - Checks that files can still be accessed on the other service
 #[serial]
 #[test]
-fn basic_text_file_transfer() {
+fn test_sending_files_on_stop() {
     start_log();
     log::info!("vvvvvv basic_text_file_transfer vvvvvv");
     let mut env = EnvironmentManager::new();
